backend/requirements.txt
alabaster==0.7.12
amqp==2.5.2
appdirs==1.4.3
argh==0.26.2
asgiref==3.2.3
attrs==19.3.0
Babel==2.8.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
codacy-coverage==1.3.11
codecov==2.0.16
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-countries==6.0
django-model-utils==4.0.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
docutils==0.16
freezegun==0.3.15
graphviz==0.13.2
gunicorn==20.0.4
idna==2.9
imagesize==1.2.0
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
livereload==2.6.1
MarkupSafe==1.1.1
more-itertools==8.2.0
oauthlib==3.1.0
packaging==20.3
pathspec==0.7.0
pathtools==0.1.2
Pillow==7.0.0
pluggy==0.13.1
port-for==0.3.1
psycopg2-binary==2.8.4
py==1.8.1
Pygments==2.6.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
sentry-sdk==0.14.2
six==1.14.0
snowballstemmer==2.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
toml==0.10.0
tornado==6.0.4
typed-ast==1.4.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
watchdog==0.10.2
wcwidth==0.1.8
zipp==3.1.0
